Refactored the LoggingExporterService that combined the export of traces and metrics into two separate classes LoggingMetricExporterService and LoggingTraceExporterService.
